requirements.txt
split-folders
pytorch
torchvision
numpy
tqdm
cv2